Go To Module with Go To Class action
New ModuleName index keeps track of only the names of modulars:
* defmodule
* defimpl
* defprotocol

It is used to power gotoClassContributor that allows using the Go To
Class action (Cmd+O) to go to modules separately from all functions as
would happen with Go To Symbols (Alt+Cmd+O).
