+++ src/org/elixir_lang/navigation/GotoClassContributor.kt
@@ -1,0 +1,5 @@
+package org.elixir_lang.navigation
+
+import org.elixir_lang.psi.stub.index.ModularName
+
+class GotoClassContributor : ChooseByNameContributor(ModularName.KEY)
