+++ src/org/elixir_lang/psi/Definition.kt
@@ -19,36 +19,50 @@ enum class Definition(val type: Type) {
     }
 }
 
+/**
+ * What kind of definition is defined by the [call]
+ */
 fun definition(call: Call): Definition? =
     call.resolvedModuleName()?.let { resolvedModuleName ->
         call.functionName()?.let { functionName ->
             if (resolvedModuleName == KERNEL) {
                 val resolvedFinalArity = call.resolvedFinalArity()
+                val hasDoBlockOrKeyword= call.hasDoBlockOrKeyword()
 
-                if (resolvedFinalArity == 2 && call.hasDoBlockOrKeyword()) {
-                    when (functionName) {
-                        DEF -> Definition.PUBLIC_FUNCTION
-                        DEFP -> Definition.PRIVATE_FUNCTION
-                        DEFMACRO -> Definition.PUBLIC_MACRO
-                        DEFMACROP -> Definition.PRIVATE_MACRO
-                        DEFMODULE -> Definition.MODULE
-                        DEFIMPL -> Definition.IMPLEMENTATION
-                        DEFPROTOCOL -> Definition.PROTOCOL
-                        else -> null
-                    }
-                } else if (resolvedFinalArity == 1) {
-                    when (functionName) {
-                        DEF -> Definition.PUBLIC_FUNCTION
-                        DEFP -> Definition.PRIVATE_FUNCTION
-                        DEFMACRO -> Definition.PUBLIC_MACRO
-                        DEFMACROP -> Definition.PRIVATE_MACRO
-                        else -> null
-                    }
-                } else {
-                    null
-                }
+                definition(functionName, resolvedFinalArity, hasDoBlockOrKeyword)
             } else {
                 null
             }
         }
     }
+
+fun definition(resolvedModuleName: String?, functionName: String?, resolvedFinalArity: Int, hasDoBlockOrKeyword: Boolean) =
+        if (resolvedModuleName == KERNEL) {
+            definition(functionName, resolvedFinalArity, hasDoBlockOrKeyword)
+        } else {
+            null
+        }
+
+fun definition(functionName: String?, resolvedFinalArity: Int, hasDoBlockOrKeyword: Boolean) =
+        if (resolvedFinalArity == 2 && hasDoBlockOrKeyword) {
+            when (functionName) {
+                DEF -> Definition.PUBLIC_FUNCTION
+                DEFP -> Definition.PRIVATE_FUNCTION
+                DEFMACRO -> Definition.PUBLIC_MACRO
+                DEFMACROP -> Definition.PRIVATE_MACRO
+                DEFMODULE -> Definition.MODULE
+                DEFIMPL -> Definition.IMPLEMENTATION
+                DEFPROTOCOL -> Definition.PROTOCOL
+                else -> null
+            }
+        } else if (resolvedFinalArity == 1) {
+            when (functionName) {
+                DEF -> Definition.PUBLIC_FUNCTION
+                DEFP -> Definition.PRIVATE_FUNCTION
+                DEFMACRO -> Definition.PUBLIC_MACRO
+                DEFMACROP -> Definition.PRIVATE_MACRO
+                else -> null
+            }
+        } else {
+            null
+        }
